+++ app/src/main/kotlin/com/hm/picplz/navigation/NavHost.kt
@@ -10,8 +10,8 @@ import androidx.navigation.NavHostController
 import com.hm.picplz.data.model.User
 import com.hm.picplz.ui.screen.main.MainScreen
 import com.hm.picplz.ui.screen.sign_up.SignUpScreen
-import com.hm.picplz.ui.screen.sign_up_client.SignUpClientScreen
-import com.hm.picplz.ui.screen.sign_up_photographer.SignUpPhotographerScreen
+import com.hm.picplz.ui.screen.sign_up.sign_up_client.SignUpClientScreen
+import com.hm.picplz.ui.screen.sign_up.sign_up_photographer.SignUpPhotographerScreen
 import com.hm.picplz.viewmodel.MainActivityUiState
 import com.hm.picplz.viewmodel.emptyUserData
