+++ app/src/main/kotlin/com/hm/picplz/ui/screen/sign_up/sign_up_client/SignUpClientIntent.kt
@@ -1,8 +1,8 @@
-package com.hm.picplz.ui.screen.sign_up_client
+package com.hm.picplz.ui.screen.sign_up.sign_up_client
 
 import android.net.Uri
 import com.hm.picplz.data.model.User
-import com.hm.picplz.ui.screen.sign_up_photographer.SignUpPhotographerIntent
+import com.hm.picplz.ui.screen.sign_up.sign_up_photographer.SignUpPhotographerIntent
 
 sealed class SignUpClientIntent {
     data class SetUserInfo(val userInfo: User) : SignUpClientIntent()
