+++ app/src/main/kotlin/com/hm/picplz/ui/screen/sign_up/sign_up_photographer/SignUpPhotographerScreen.kt
@@ -1,6 +1,5 @@
-package com.hm.picplz.ui.screen.sign_up_client
+package com.hm.picplz.ui.screen.sign_up.sign_up_photographer
 
-import SignUpClientNicknameView
 import android.widget.Toast
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Scaffold
@@ -13,31 +12,31 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import com.hm.picplz.data.model.User
 import com.hm.picplz.ui.theme.MainThemeColor
-import com.hm.picplz.viewmodel.SignUpClientViewModel
+import com.hm.picplz.viewmodel.SignUpPhotographerViewModel
 import com.hm.picplz.viewmodel.emptyUserData
 import kotlinx.coroutines.flow.collectLatest
 
 @Composable
-fun SignUpClientScreen(
+fun SignUpPhotographerScreen(
     modifier: Modifier = Modifier,
     navController: NavController,
     userInfo: User = emptyUserData,
-    viewModel: SignUpClientViewModel = viewModel(),
+    viewModel: SignUpPhotographerViewModel = viewModel(),
 ) {
     val currentState = viewModel.state.collectAsState().value
 
-    Scaffold(
+    Scaffold (
         modifier = Modifier.fillMaxSize(),
         containerColor = MainThemeColor.White
     ) { innerPadding ->
         when (currentState.currentStep) {
-            0 -> SignUpClientNicknameView(
+            0 -> SignUpPhotographerNicknameView(
                 modifier = modifier,
                 currentState = currentState,
                 viewModel = viewModel,
-                innerPadding = innerPadding
+                innerPadding = innerPadding,
             )
-            1 -> SignUpClientProfileImageView(
+            1 -> SignUpPhotographerProfileImageView(
                 modifier = modifier,
                 currentState = currentState,
                 viewModel = viewModel,
@@ -51,12 +50,14 @@ fun SignUpClientScreen(
     LaunchedEffect(Unit) {
         viewModel.sideEffect.collectLatest { sideEffect ->
             when (sideEffect) {
-                is SignUpClientSideEffect.SubmitProfileInfo -> {
-                    Toast.makeText(context, "가입", Toast.LENGTH_SHORT).show()
-                }
-                is SignUpClientSideEffect.ShowFileUploadDialog -> {}
-                is SignUpClientSideEffect.NavigateToPrev-> {
+                is SignUpPhotographerSideEffect.NavigateToPrev -> {
                     navController.popBackStack()
+                }
+                is SignUpPhotographerSideEffect.ShowFileUploadDialog -> {
+
+                }
+                is SignUpPhotographerSideEffect.SubmitProfileInfo -> {
+                    Toast.makeText(context, "가입", Toast.LENGTH_SHORT).show()
                 }
             }
         }
